first implementation of karma bot 4. only functionality at the moment is reposting
