+++ Pipfile
@@ -1,0 +1,13 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+psaw = "*"
+praw = "*"
+
+[requires]
+python_version = "3.9"
