--- requirements.txt
@@ -1,1 +1,0 @@
--r src/requirements.txt
