+++ run_linux.sh
@@ -2,6 +2,7 @@
 
 DEBUG_FILE="./run_linux.log"
 SETTINGS_FILE="./src/settings.py"
+export PIPENV_VENV_IN_PROJECT=1
 
 date '+%d/%m/%Y %H:%M:%S' | tee $DEBUG_FILE
 
@@ -27,10 +28,9 @@ case "${unameOut}" in
 esac
 echo "system is ${machine}" | tee -a $DEBUG_FILE
 
-
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
-if [ ! -d "$DIR/venv" ]; then
+if [ ! -d "$DIR/.venv" ]; then
   echo "no virtualenv detected doing setup before running" | tee -a $DEBUG_FILE
   echo "need to install dependencies" | tee -a $DEBUG_FILE
   if [ "${machine}" =  "Linux" ]; then
@@ -40,39 +40,30 @@ if [ ! -d "$DIR/venv" ]; then
     apt-get install -y --no-install-recommends \
       g++ \
       gcc \
-      golang-go \
       libc6-dev \
       make \
       pkg-config \
       libffi-dev \
-      wget \
-      tmux \
       python3.6 \
       python3-pip \
       python3-setuptools \
       python3-dev \
       git || { echo 'Installing dependencies failed' | tee -a $DEBUG_FILE ; exit 1; }
   elif [ "${machine}" =  "Mac" ]; then
-    xcode-select --install
+    $(xcode-select -p) || xcode-select --install
   else
     echo "No suitable linux version" | tee -a $DEBUG_FILE
   fi
 
   if [ "${machine}" =  "Linux" ] || [ "${machine}" =  "Mac" ]; then
-    pip3 install virtualenv || { echo 'Installing virtualenv failed' | tee -a $DEBUG_FILE ; exit 1; }
-    virtualenv -p "$(command -v python3)" venv || { echo 'Installing virtualenv failed' | tee -a $DEBUG_FILE ; exit 1; }
-    source venv/bin/activate
-    pip3 install cython
-    pip3 install -r ./src/requirements.txt || { echo 'Installing python deps failed' | tee -a $DEBUG_FILE ; exit 1; }
+    pip3 install pipenv || { echo 'Installing virtualenv failed' | tee -a $DEBUG_FILE ; exit 1; }
+    pipenv install || { echo 'Installing python deps failed' | tee -a $DEBUG_FILE ; exit 1; }
   fi
-
 fi
 
-
-
-echo "activating the virtualenv " | tee -a $DEBUG_FILE
-source venv/bin/activate
-
 echo "trying to run the bot" | tee -a $DEBUG_FILE
-python3 ./src/run.py "$@"
+pipenv run python3 ./src/init.py "$@"
+
+
+
