+++ src/actions/__init__.py
@@ -1,0 +1,1 @@
+from .action import Action
