+++ src/actions/action.py
@@ -1,0 +1,10 @@
+class Action():
+  def __init__(self):
+    self.ready = False
+
+  def _init(self):
+    self.ready = True
+
+  def action(self):
+    self._init()
+    pass
