+++ src/bots/reddit/__init__.py
@@ -1,0 +1,2 @@
+from .bot import RedditBot
+
