+++ src/bots/reddit/actions/posts.py
@@ -1,0 +1,9 @@
+from apis.pushshift import PushShift
+from actions import Action
+
+class Posts(Action):
+  def __init__(self):
+    self.psapi = PushShift()
+
+  def action(self):
+    self.psapi.get_posts("pcmasterrace")
