+++ src/init.py
@@ -1,4 +1,4 @@
-import init
+import bot
 
 if __name__ == "__main__":
-  init.init()
+  bot.run()
