--- src/requirements.txt
@@ -1,79 +1,0 @@
-appdirs==1.4.4
-attrs==19.3.0
-backports.functools-lru-cache==1.6.1
-beautifulsoup4==4.9.0
-BTrees==4.7.2
-certifi==2020.4.5.1
-cffi==1.14.0
-chardet==3.0.4
-click==7.1.2
-cobe==3.0.0
-configparser==5.0.0
-contextlib2==0.6.0.post1
-Cython==0.29.17
-distlib==0.3.0
-filelock==3.0.12
-Flask==1.1.2
-idna==2.9
-imgurpython==1.1.7
-importlib-metadata==1.6.0
-importlib-resources==1.5.0
-inflect==4.1.0
-irc==18.0.0
-itsdangerous==1.1.0
-jaraco.classes==3.1.0
-jaraco.collections==3.0.0
-jaraco.functools==3.0.1
-jaraco.itertools==5.0.0
-jaraco.logging==3.0.0
-jaraco.stream==3.0.0
-jaraco.text==3.2.0
-Jinja2==2.11.2
-joblib==0.14.1
-MarkupSafe==1.1.1
-memory-profiler==0.57.0
-more-itertools==8.2.0
-mypy==0.770
-mypy-extensions==0.4.3
-nltk==3.5
-packaging==20.4
-pathlib2==2.3.5
-persistent==4.6.4
-pipenv==2020.6.2
-pluggy==0.13.1
-praw==7.0.0
-prawcore==1.3.0
-psutil==5.7.0
-py==1.8.1
-pycparser==2.20
-pyparsing==2.4.7
-PyStemmer==2.0.0.1
-pytest==5.4.1
-pytest-mypy==0.4.2
-pytz==2020.1
-regex==2020.4.4
-requests==2.23.0
-scandir==1.10.0
-singledispatch==3.4.0.3
-six==1.14.0
-soupsieve==2.0
-tempora==3.0.0
-tinydb==4.0.0.post1
-tqdm==4.46.0
-transaction==3.0.0
-typed-ast==1.4.1
-typing==3.7.4.1
-typing-extensions==3.7.4.2
-update-checker==0.17
-urllib3==1.25.9
-virtualenv==20.0.21
-virtualenv-clone==0.5.4
-wcwidth==0.1.9
-websocket-client==0.57.0
-Werkzeug==1.0.1
-zc.lockfile==2.0
-ZConfig==3.5.0
-zipp==3.1.0
-ZODB==5.5.1
-zodbpickle==2.0.0
-zope.interface==5.1.0
