+++ src/run.sh
@@ -5,4 +5,4 @@ mkdir -p /reddit-karma-bot-run
 cp -rf /reddit-karma-bot-src/* /reddit-karma-bot-run || true
 mkdir -p /reddit-kamra-bot-run/brains || true
 mkdir -p /reddit-karma-bot/brains || true
-python3 /reddit-karma-bot-run/run.py
+python3 /reddit-karma-bot-run/init.py
